+++ package.json
@@ -37,7 +37,7 @@
     "react-native-vector-icons": "^9.2.0",
     "react-native-web": "0.18.10",
     "react-spring-bottom-sheet": "^3.4.1",
-    "spelieve-common": "^1.1.75"
+    "spelieve-common": "^1.1.77"
   },
   "devDependencies": {
     "@babel/core": "7.19.6",
